Refactor sequence match process
Previously, sequences were checked, after checking the possible
vim matches, against a serious of possible partial matches. This
was unmaintainable, since it was not clear, as we were adding more
vim regex matches, which matches were already included in the partial match
regexes and which needed to be added; it was also difficult to be
certain that these partial match checks protected against invalid
sequences being entered.